[SASS-11423] Small refactors following review comments
