+++ app/views/pages/businessTaxReliefs/CheckAnswersQualifyingLoanPageView.scala.html
@@ -15,9 +15,9 @@
  *@
 
 @import models.requests.JourneyDataRequest
-@import views.html.templates.Layout
-@import views.html.templates.helpers.{Button, ErrorSummary, Heading, InputText, Link}
 @import viewmodels.businessTaxReliefs.CheckAnswersQualifyingLoanSummary
+@import views.html.templates.Layout
+@import views.html.templates.helpers.{Button, Heading}
 
 @this(
         layout: Layout,
